+++ src/events/onceReady.js
@@ -32,7 +32,7 @@ module.exports = class extends Event {
 		// Added for consistency with other datastores, Client#clients does not exist
 		clientStorage.cache.set(this.client.user.id, this.client);
 		this.client.settings = clientStorage.create(this.client, this.client.user.id);
-		await this.client.gateways.sync();
+		await Promise.all(this.client.gateways.map(gateway => gateway.sync()));
 
 		// Init the schedule
 		await this.client.schedule.init();